pyproject.toml
[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[project]
name = "mypycli"
description = "Python framework for building extensible CLI applications with modules, REPL console, and daemon mode."
requires-python = ">=3.10"
authors = [
    { name = "nessshon" },
]
maintainers = [
    { name = "nessshon" },
]
license = "MIT"
readme = "README.md"
dynamic = ["version"]
keywords = [
    "CLI",
    "console",
    "daemon",
    "framework",
    "modules",
    "REPL",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0",
    "psutil==7.0.0",
    "requests>=2.31.0",
    "gitpython>=3.1",
    "packaging>=24.0",
    "types-psutil",
    "types-requests",
    "pyyaml>=6",
    "types-PyYAML",
]

[project.scripts]
mypycli = "mypycli.cli.standalone:main"

[project.urls]
Homepage = "https://github.com/nessshon/mypycli"

[tool.setuptools.dynamic]
version = { attr = "mypycli.__meta__.__version__" }

[tool.setuptools.packages.find]
include = ["mypycli", "mypycli.*"]

[tool.setuptools.package-data]
mypycli = ["py.typed", "i18n/locales/*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # isort
    "I",
    # flake8-bugbear
    "B",
    # flake8-comprehensions
    "C4",
    # pyupgrade
    "UP",
    # flake8-simplify
    "SIM",
    # type-checking imports
    "TCH",
    # perflint
    "PERF",
    # flake8-debugger
    "T10",
    # flake8-print
    "T20",
    # flake8-pie
    "PIE",
    # pylint errors
    "PLE",
    # flake8-return
    "RET",
    # ruff-specific
    "RUF",
    # flake8-bandit (security)
    "S",
]
ignore = [
    # no module-level docstrings
    "D100",
    # blank line before class docstring (conflicts D211)
    "D203",
    # multi-line summary second line (conflicts D212)
    "D213",
    # print (library intentionally uses print)
    "T201",
    # subprocess with untrusted input (library intentionally calls system commands)
    "S603",
    # os.system shell injection (library intentionally uses shell commands)
    "S605",
    # subprocess without full path (paths differ across OS)
    "S607",
    # missing docstring in __init__
    "D107",
    # missing docstring in public method (for self-explanatory one-liners)
    "D102",
    # missing docstring in public function
    "D103",
    # insecure /tmp usage (intentional for temp files)
    "S108",
    # missing docstring in public class (enums, dataclasses)
    "D101",
    # missing docstring in magic method
    "D105",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"tests/**" = ["S101", "S104", "S105", "S106", "B017", "D", "RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["mypycli"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
strict = true
pretty = true
show_error_codes = true
show_error_context = true
namespace_packages = true
follow_imports_for_stubs = true
warn_unreachable = true
enable_error_code = [
    "ignore-without-code",
    "redundant-cast",
    "truthy-bool",
]
disable_error_code = [
    "type-abstract",
]